+++ pkg/_fe_analyzer_shared/lib/src/parser/listener.dart
@@ -1692,6 +1692,10 @@ class Listener implements UnescapeErrorListener {
     logEvent("NamedArgument");
   }
 
+  void handleNamedRecordField(Token colon) {
+    logEvent("NamedRecordField");
+  }
+
   void beginNewExpression(Token token) {}
 
   void endNewExpression(Token token) {
@@ -1746,10 +1750,19 @@ class Listener implements UnescapeErrorListener {
     logEvent("ParenthesizedCondition");
   }
 
-  /// Handle a parenthesized expression.
+  /// Starts a parenthesized expression or a record literal. Will be ended with
+  /// either [endParenthesizedExpression] or [endRecordLiteral].
+  void beginParenthesizedExpressionOrRecordLiteral(Token token) {}
+
+  /// Ends a record literal with [count] entries.
+  void endRecordLiteral(Token token, int count) {
+    logEvent("RecordLiteral");
+  }
+
+  /// End a parenthesized expression.
   /// These may be within the condition expression of a control structure
   /// but will not be the condition of a control structure.
-  void handleParenthesizedExpression(Token token) {
+  void endParenthesizedExpression(Token token) {
     logEvent("ParenthesizedExpression");
   }