+++ pkg/dart2wasm/lib/constants.dart
@@ -83,8 +83,8 @@ class Constants {
       b.i32_const(info.classId);
       b.i32_const(initialIdentityHash);
       b.i32_const(0);
-      translator.array_new_default(b, arrayType);
-      translator.struct_new(b, info);
+      b.array_new_default(arrayType);
+      b.struct_new(info.struct);
       b.global_set(emptyString);
     } else {
       w.RefType emptyStringType = info.nonNullableType;
@@ -92,8 +92,8 @@ class Constants {
       w.Instructions ib = emptyString.initializer;
       ib.i32_const(info.classId);
       ib.i32_const(initialIdentityHash);
-      translator.array_init(ib, arrayType, 0);
-      translator.struct_new(ib, info);
+      ib.array_new_fixed(arrayType, 0);
+      ib.struct_new(info.struct);
       ib.end();
     }
 
@@ -121,8 +121,8 @@ class Constants {
       b.ref_null(typeInfo.struct); // Initialized later
       b.i64_const(0);
       b.i32_const(0);
-      translator.array_new_default(b, arrayType);
-      translator.struct_new(b, info);
+      b.array_new_default(arrayType);
+      b.struct_new(info.struct);
       b.global_set(emptyTypeList);
     } else {
       w.RefType emptyListType = info.nonNullableType;
@@ -132,8 +132,8 @@ class Constants {
       ib.i32_const(initialIdentityHash);
       ib.ref_null(typeInfo.struct); // Initialized later
       ib.i64_const(0);
-      translator.array_init(ib, arrayType, 0);
-      translator.struct_new(ib, info);
+      ib.array_new_fixed(arrayType, 0);
+      ib.struct_new(info.struct);
       ib.end();
     }
 
@@ -199,7 +199,7 @@ class Constants {
   /// in the corresponding string data segment from which to copy this string.
   w.DefinedFunction makeStringFunction(Class cls) {
     ClassInfo info = translator.classInfo[cls]!;
-    w.FunctionType ftype = translator.functionType(
+    w.FunctionType ftype = m.addFunctionType(
         const [w.NumType.i32, w.NumType.i32], [info.nonNullableType]);
     return m.addFunction(ftype, "makeString (${cls.name})");
   }
@@ -218,7 +218,7 @@ class Constants {
 
     w.Instructions b = function.body;
     b.local_get(length);
-    translator.array_new_default(b, arrayType);
+    b.array_new_default(arrayType);
     b.local_set(array);
 
     b.i32_const(0);
@@ -243,7 +243,7 @@ class Constants {
     b.i32_const(info.classId);
     b.i32_const(initialIdentityHash);
     b.local_get(array);
-    translator.struct_new(b, info);
+    b.struct_new(info.struct);
     b.end();
   }
 
@@ -359,7 +359,7 @@ class ConstantInstantiator extends ConstantVisitor<w.ValueType> {
       ClassInfo info = translator.classInfo[translator.boxedClasses[wasmType]]!;
       b.i32_const(info.classId);
       pushValue();
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
       return info.nonNullableType;
     } else {
       pushValue();
@@ -419,7 +419,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           m.addGlobal(w.GlobalType(type.withNullability(true)));
       global.initializer.ref_null(type.heapType);
       global.initializer.end();
-      w.FunctionType ftype = translator.functionType(const [], [type]);
+      w.FunctionType ftype = m.addFunctionType(const [], [type]);
       w.DefinedFunction function = m.addFunction(ftype, "$constant");
       generator(function, function.body);
       w.Local temp = function.addLocal(type);
@@ -493,15 +493,15 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           segment.append(bytes);
           b.i32_const(offset);
           b.i32_const(constant.value.length);
-          translator.array_init_from_data(b, arrayType, segment);
+          b.array_new_data(arrayType, segment);
         } else {
           // Initialize string contents from i32 constants on the stack.
           for (int charCode in constant.value.codeUnits) {
             b.i32_const(charCode);
           }
-          translator.array_init(b, arrayType, constant.value.length);
+          b.array_new_fixed(arrayType, constant.value.length);
         }
-        translator.struct_new(b, info);
+        b.struct_new(info.struct);
       }
     });
   }
@@ -551,7 +551,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
         constants.instantiateConstant(
             function, b, subConstant, info.struct.fields[i].type.unpacked);
       }
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -580,7 +580,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
         // Allocate array and set each entry to the corresponding sub-constant.
         w.Local arrayLocal = function!.addLocal(refType.withNullability(false));
         b.i32_const(length);
-        translator.array_new_default(b, arrayType);
+        b.array_new_default(arrayType);
         b.local_set(arrayLocal);
         for (int i = 0; i < length; i++) {
           b.local_get(arrayLocal);
@@ -596,9 +596,9 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           constants.instantiateConstant(
               function, b, constant.entries[i], elementType);
         }
-        translator.array_init(b, arrayType, length);
+        b.array_new_fixed(arrayType, length);
       }
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -636,7 +636,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           function, b, keyTypeConstant, constants.typeInfo.nullableType);
       constants.instantiateConstant(
           function, b, valueTypeConstant, constants.typeInfo.nullableType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -665,7 +665,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
       b.i64_const(0); // _deletedKeys
       constants.instantiateConstant(
           function, b, elementTypeConstant, constants.typeInfo.nullableType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -700,7 +700,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
       } else {
         b.ref_func(closureFunction);
       }
-      translator.struct_new(b, parameterCount);
+      b.struct_new(translator.closureStructType(parameterCount));
     });
   }
 
@@ -719,7 +719,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
       b.i64_const(typeInfo.classId);
       constants.instantiateConstant(
           function, b, typeArgs, typeListExpectedType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -733,7 +733,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
       types.encodeNullability(b, type);
       constants.instantiateConstant(
           function, b, typeArgument, types.nonNullableTypeType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -763,7 +763,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           function, b, requiredParameterCountConstant, w.NumType.i64);
       constants.instantiateConstant(function, b, namedParametersConstant,
           types.namedParametersExpectedType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 
@@ -785,7 +785,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
           b.i32_const(info.classId);
           b.i32_const(initialIdentityHash);
           types.encodeNullability(b, type);
-          translator.struct_new(b, info);
+          b.struct_new(info.struct);
         });
       } else {
         return _makeFunctionType(constant, type, info);
@@ -806,7 +806,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
         // explicitly declared to be nullabe.
         b.i32_const(type.declaredNullability == Nullability.nullable ? 1 : 0);
         b.i64_const(environmentIndex);
-        translator.struct_new(b, info);
+        b.struct_new(info.struct);
       });
     } else {
       assert(type is VoidType ||
@@ -817,7 +817,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
         b.i32_const(info.classId);
         b.i32_const(initialIdentityHash);
         types.encodeNullability(b, type);
-        translator.struct_new(b, info);
+        b.struct_new(info.struct);
       });
     }
   }
@@ -834,7 +834,7 @@ class ConstantCreator extends ConstantVisitor<ConstantInfo?> {
       b.i32_const(info.classId);
       b.i32_const(initialIdentityHash);
       constants.instantiateConstant(function, b, nameConstant, stringType);
-      translator.struct_new(b, info);
+      b.struct_new(info.struct);
     });
   }
 }